backend/common/requirements.txt
alembic==1.14.1
greenlet==3.1.1
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
